requirements.txt
# requirements.txt
torch
transformers
Pillow
psycopg2-binary
pandas
numpy
fastapi
uvicorn[standard]
python-multipart